Rewrite the actuator controller with actual values.
We need to use TIME, instead of DISTANCE.
We know how far it moves from end to end, and we know the resistance
of those points AND we know how long it takes to move.
From that, we can calculate how long to enable the actuator to move
a certain distance.

Calculate the current gear the actuator is in, based on the resistance
of the potentiometer.